refactor(form): extracted Form component from index and added link for fonts
