+++ src/components/Deck.js
@@ -32,6 +32,9 @@ export class Deck extends PureComponent {
     })
   }
 
+  /**
+   * HACK: There is likely a better way to traverse through cards
+   */
   nextCard = () => {
     this.setState({
       counts: {
@@ -43,6 +46,9 @@ export class Deck extends PureComponent {
     })
   }
 
+  /**
+   * HACK: There is likely a better way to traverse through cards
+   */
   prevCard = () => {
     this.setState({
       counts: {
