+++ src/index.css
@@ -1,7 +1,5 @@
-@import url('https://fonts.googleapis.com/css?family=Source+Sans+Pro');
-
-* {
-  font-family: 'Source Sans Pro', sans-serif;
+body {
+  font-family: 'Titillium Web', sans-serif;
   text-align: center;
 }
